refactor: isolate usage normalization dependencies
